Removed unnecessary files, moved files and changed routes accordingly
