--- src/redux/band/action.type.js
@@ -1,7 +1,0 @@
-const BandActionTypes = {
-    ACTIVITY:"ACTIVITY",
-    ACTIVITY_SUCCESS:"ACTVITY_SUCCESS",
-    ACTIVITY_FAILURE:"ACTVITY_FAILURE"
-};
-
-export default BandActionTypes;
