+++ src/redux/band/reducer.js
@@ -1,22 +1,21 @@
-import BandActionTypes from './action.type';
-//import BandZipActions from './zip/action.type';
+import BandZipActions from './zip/action.type';
 
 const initState = {
-    isLoading:false,
-    errorMessage:'',
+    isLoading: false,
+    errorMessage: ""
 }
 
-const activityReducer = (state=initState, action) => {
-    switch(action.type) {
-        case BandActionTypes.ACTIVITY:
-            return {isLoading:true, errorMessage:''};
-        case BandActionTypes.ACTIVITY_SUCCESS:
-            return {isLoading:false, errorMessage:''};
-        case BandActionTypes.ACTIVITY_FAILURE:
-            return {isLoading:false, errorMessage:action.payload}
+const zipReducer = (state = initState, action) => {
+    switch (action.type) {
+        case BandZipActions.UPLOAD:
+            return { isLoading: true, errorMessage: "" };
+        case BandZipActions.UPLOAD_SUCCESS:
+            return { isLoading: false, errorMessage: "" };
+        case BandZipActions.UPLOAD_FAILURE:
+            return { isLoading: false, errorMessage: action.payload };
         default:
             return state;
     }
 }
 
-export default activityReducer;
+export default zipReducer;
