+++ src/redux/saga.js
@@ -2,7 +2,7 @@ import { all } from "redux-saga/effects";
 import AuthSaga from "redux/auth/saga";
 import UserSaga from "redux/users/saga";
 import BandSaga from "redux/band/saga";
-import UploadSaga from "redux/band/zip/saga";
+import UploadSaga from "redux/band/saga";
 const rootSaga = function* () {
   yield all([
     AuthSaga(),
